+++ saleor/graphql/product/tests/test_product.py
@@ -950,6 +950,21 @@ def test_products_query_with_filter_stocks(
     assert {node["node"]["id"] for node in products_data} == product_ids
 
 
+def test_query_products_query_with_filter_ids(
+    staff_api_client, product, query_products_with_filter
+):
+    product_global_id = graphene.Node.to_global_id("Product", product.id)
+    variables = {"filter": {"ids": [product_global_id]}}
+    response = staff_api_client.post_graphql(
+        query_products_with_filter, variables, check_no_permissions=False
+    )
+    content = get_graphql_content(response)
+    products_data = content["data"]["products"]["edges"]
+
+    assert len(products_data) == 1
+    assert products_data[0]["node"]["id"] == product_global_id
+
+
 def test_query_product_image_by_id(user_api_client, product_with_image):
     image = product_with_image.images.first()
     query = """